Moved more stuff over to tablefq.T
