+++ libsq/core/tablefq/tablefq.go
@@ -29,29 +29,31 @@ func (t T) String() string {
 	return t.Render(stringz.DoubleQuote)
 }
 
-// Render renders t using escapeFn, e.g. stringz.DoubleQuote.
-func (t T) Render(escapeFn func(s string) string) string {
+// Render renders t using quoteFn, e.g. stringz.DoubleQuote.
+func (t T) Render(quoteFn func(s string) string) string {
 	sb := strings.Builder{}
 	if t.Catalog != "" {
-		sb.WriteString(escapeFn(t.Catalog))
+		sb.WriteString(quoteFn(t.Catalog))
 		sb.WriteRune('.')
 	}
 	if t.Schema != "" {
-		sb.WriteString(escapeFn(t.Schema))
+		sb.WriteString(quoteFn(t.Schema))
 		sb.WriteRune('.')
 	}
-	sb.WriteString(escapeFn(t.Table))
+	sb.WriteString(quoteFn(t.Table))
 	return sb.String()
 }
 
-func New(name string) T {
-	return T{Table: name}
+// New returns a new tablefq.T with the given T.Table value,
+// and empty T.Catalog and T.Schema values.
+func New(tbl string) T {
+	return T{Table: tbl}
 }
 
-// From returns a T from a Table, which can be either
+// From returns a T from an Any, which can be either
 // a string (just the table name) or a fully-qualified table,
 // including catalog, schema, and table name.
-func From[X Table](t X) T {
+func From[X Any](t X) T {
 	if tfq, ok := any(t).(T); ok {
 		return tfq
 	}
@@ -63,8 +65,10 @@ func From[X Table](t X) T {
 	panic(fmt.Sprintf("invalid type %T: %v", t, t))
 }
 
-// Table is a type constraint for a table name, which can be a string (just
+// Any is a type constraint for a table name, which can be a string (just
 // the table name without catalog or schema) or a fully-qualified T.
-type Table interface {
+//
+// REVISIT: Probably get rid of tablefq.Any, if it's only used in one place.
+type Any interface {
 	string | T
 }
